feat: :sparkles: Add update to saleService
It receives the id of the sale to be updated and the product list. First the product list data is validated. Next, the sale id is validated. Then, the rows from sale_product table related to the sale are removed. Finaly, new sale informations are added to the sale_product table.