docs: :memo: Uses swagger to generate the documentation and link it to the /doc route
The Swagger auto generates the documentation to and .json file. It is used by the /doc router to presente the documentation using the swaggerUi server. However, by now, the documentation does not include the sale routes, and many informations need to be include in the swagger_outpu.json file to improve the documentation.